requirements.txt
# Web Framework & Database
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0

# AI Scripting
google-generativeai==0.3.2

# AI Voice (Text-to-Speech)
edge-tts==7.2.7

# Video Processing & Editing
moviepy==1.0.3
Pillow==9.5.0
numpy==1.26.3

# API Requests (For Pexels/Visuals)
requests==2.31.0

# Utilities
aiofiles==23.2.1
asyncio==3.4.3